ADR #5: Persistency for LocalBlobAccess
Even though LocalBlobAccess has the potential to replace
CircularBlobAccess, it currently lacks a couple of features that are
pretty important. There is no support for file-backed storage, nor can
data be persisted across restarts.

Because adding persistency is going to be pretty complex, let's release
an ADR that acts as a reference for how all of this is going to work.